Move data to testdata, split MultiDBSnapshot from MultiDBState
