--- storagenode/storagenodedb/storagenodedbtest/v13.go
@@ -1,9 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package storagenodedbtest
-
-var v13 = MultiDBSnapshot{
-	Version:   13,
-	Databases: v12.Databases,
-}
