--- storagenode/storagenodedb/storagenodedbtest/v14.go
@@ -1,9 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package storagenodedbtest
-
-var v14 = MultiDBSnapshot{
-	Version:   14,
-	Databases: v13.Databases,
-}
