--- storagenode/storagenodedb/storagenodedbtest/v5.go
@@ -1,9 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package storagenodedbtest
-
-var v5 = MultiDBSnapshot{
-	Version:   5,
-	Databases: v4.Databases,
-}
