+++ storagenode/storagenodedb/testdata/v13.go
@@ -1,0 +1,9 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+package testdata
+
+var v13 = MultiDBState{
+	Version:  13,
+	DBStates: v12.DBStates,
+}
