+++ storagenode/storagenodedb/testdata/v14.go
@@ -1,0 +1,9 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+package testdata
+
+var v14 = MultiDBState{
+	Version:  14,
+	DBStates: v13.DBStates,
+}
