+++ storagenode/storagenodedb/testdata/v5.go
@@ -1,0 +1,9 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+package testdata
+
+var v5 = MultiDBState{
+	Version:  5,
+	DBStates: v4.DBStates,
+}
